refactor: move userDropdown to custom file
